// src/document/types/document.types.ts
export enum DocumentType {
  ADVANCE_PAYMENT = "ADVANCE_PAYMENT", // ADIANTAMENTOS
  BANK_SLIP = "BANK_SLIP", // BOLETO
  CHECK = "CHECK", // CHEQUE
  CREDIT_CARD = "CREDIT_CARD", // CARTÃO DE CRÉDITO
  POST_DATED_CHECK = "POST_DATED_CHECK", // CHEQUE PRÉ DATADO
  TRADE_BILL = "TRADE_BILL", // DUPLICATA
  LOAN = "LOAN", // EMPRÉSTIMO
  PETTY_CASH = "PETTY_CASH", // FUNDO DE CAIXA
  VACATION_PAY = "VACATION_PAY", // FÉRIAS
  SEVERANCE_FUND = "SEVERANCE_FUND", // FGTS
  PAYROLL = "PAYROLL", // FOLHA
  SOCIAL_SECURITY = "SOCIAL_SECURITY", // INSS
  TAXES = "TAXES", // IMPOSTOS
  CREDIT_NOTE = "CREDIT_NOTE", // NOTA DE CRÉDITO
  SUPPLIER_RETURN_NOTE = "SUPPLIER_RETURN_NOTE", // NOTA DE DEVOLUÇÃO FORNECEDOR
  ADVANCE_PAYMENT_TO_SUPPLIER = "ADVANCE_PAYMENT_TO_SUPPLIER", // PAGAMENTO ANTECIPADO
  BRADESCO_HEALTH_PLAN = "BRADESCO_HEALTH_PLAN", // PLANO DE SAÚDE BRADESCO
  FORECAST = "FORECAST", // PREVISÃO
  ADVANCE_RECEIPT = "ADVANCE_RECEIPT", // RECEBIMENTO ANTECIPADO
  RECEIPT = "RECEIPT", // RECIBO
  TERMINATION = "TERMINATION", // RESCISÃO
  HEALTH_PLAN = "HEALTH_PLAN", // PLANO DE SAÚDE
  TRIBUTES = "TRIBUTES", // TRIBUTOS
  FEES = "FEES", // TAXAS
  GENERAL = "GENERAL", // GERAL
}

export enum DocumentStatus {
  PENDING = "PENDING",
  IN_PROGRESS = "IN_PROGRESS",
  COMPLETED = "COMPLETED",
  CANCELLED = "CANCELLED",
  EXPIRED = "EXPIRED",
  DRAFT = "DRAFT",
  REJECTED = "REJECTED",
}
// ...
export interface IDocument {
  id: number;
  title: string;
  description: string | null;
  tipoDocumento: DocumentType;
  fileId: number;
  natureza: string;
  status: DocumentStatus;
  valor: number | null;
  dataVencimento: Date | null;
  observacoes: string | null;
  createdAt: Date;
  updatedAt: Date;
  deletedAt?: Date;
  ownerId: number | null;
  fornecedorId: number | null;
}